backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
google-generativeai>=0.8.0
Pillow>=11.0.0
python-multipart>=0.0.18
aiofiles>=24.1.0
python-dotenv>=1.0.0
